requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.2
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
dill==0.3.8
distro==1.9.0
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-openai==0.1.10
langchain-text-splitters==0.2.1
langsmith==0.1.81
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.31.0
orjson==3.10.5
packaging==24.1
platformdirs==4.2.2
pydantic==2.7.3
pydantic_core==2.18.4
pylint==3.2.4
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.4.1
tiktoken==0.7.0
tomlkit==0.12.5
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.1
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
youtube-transcript-api==0.6.2
